* Create a base class which contains the common functions for all Cryosparc protocols.
