+++ CHANGES.txt
@@ -1,3 +1,5 @@
-Dec 9, 2019 - Migrated to python3
-May 27, 2019 - Initial plugin version
+Apr 10, 2020  - Create a protocol base, Clean the code
+Apr  4, 2020  - Add the use of multiple GPUs
+Dec  9, 2019  - Migrated to python3
+May 27, 2019  - Initial plugin version
