+++ cryosparc2/protocols/__init__.py
@@ -24,7 +24,7 @@
 # *  e-mail address 'scipion@cnb.csic.es'
 # *
 # **************************************************************************
-
+from .protocol_base import ProtCryosparcBase
 from .protocol_cryorefine import ProtCryoSparcRefine3D
 from .protocol_cryosparc2d import ProtCryo2D
 from .protocol_cryosparc_ab import ProtCryoSparcInitialModel
