+++ cryosparc2/protocols/protocol_cryorefine.py
@@ -30,15 +30,13 @@ from pyworkflow.protocol.params import (PointerParam, FloatParam,
 from pwem.objects import Volume, FSC
 from pwem.protocols import ProtRefine3D
 
+from . import ProtCryosparcBase
 from ..convert import *
 from ..utils import *
 from ..constants import *
 
-import os
-import ast
 
-
-class ProtCryoSparcRefine3D(ProtRefine3D):
+class ProtCryoSparcRefine3D(ProtCryosparcBase, ProtRefine3D):
     """ Protocol to refine a 3D map using cryosparc.
         Rapidly refine a single homogeneous structure to high-resolution and
         validate using the gold-standard FSC.
@@ -62,7 +60,7 @@ class ProtCryoSparcRefine3D(ProtRefine3D):
                       label="Input particles", important=True,
                       validators=[Positive],
                       help='Select the input images from the project.')
-        form.addParam('referenceVolume', PointerParam, pointerClass='Volume',
+        form.addParam('refVolume', PointerParam, pointerClass='Volume',
                       important=True,
                       label="Input volume",
                       help='Initial reference 3D map, it should have the same '
@@ -309,41 +307,9 @@ class ProtCryoSparcRefine3D(ProtRefine3D):
         self._insertFunctionStep('createOutputStep')
 
     # --------------------------- STEPS functions ------------------------------
-    def convertInputStep(self):
-        """ Create the input file in STAR format as expected by Relion.
-        If the input particles comes from Relion, just link the file.
-        """
-        imgSet = self._getInputParticles()
-        # Create links to binary files and write the relion .star file
-        writeSetOfParticles(imgSet, self._getFileName('input_particles'),
-                            self._getTmpPath())
-
-        self._importParticles()
-
-        self.vol_fn = os.path.join(os.getcwd(),
-                                   convertBinaryVol(self.referenceVolume.get(),
-                                                    self._getTmpPath()))
-        self.importVolume = doImportVolumes(self, self.vol_fn, 'map',
-                                            'Importing volume...')
-        self.currenJob.set(self.importVolume.get())
-        self._store(self)
-
-        if self.refMask.get() is not None:
-            self.maskFn = os.path.join(os.getcwd(),
-                                       convertBinaryVol(self.refMask.get(),
-                                                        self._getTmpPath()))
-
-            self.importMask = doImportVolumes(self, self.maskFn, 'mask',
-                                              'Importing mask... ')
-            self.currenJob.set(self.importMask.get())
-            self._store(self)
-            self.mask = self.importMask.get() + '.imported_mask.mask'
-        else:
-            self.mask = None
-
     def processStep(self):
         self.vol = self.importVolume.get() + '.imported_volume.map'
-        print("Refinement started...")
+        print(pwutils.greenStr("Refinement started..."))
         self.doRunRefine()
 
     def createOutputStep(self):
@@ -448,13 +414,6 @@ class ProtCryoSparcRefine3D(ProtRefine3D):
         self._defineOutputs(outputFSC=fsc)
         self._defineSourceRelation(vol, fsc)
 
-    def setAborted(self):
-        """ Set the status to aborted and updated the endTime. """
-        ProtRefine3D.setAborted(self)
-        killJob(str(self.projectName.get()), str(self.currenJob.get()))
-        clearJob(str(self.projectName.get()), str(self.currenJob.get()))
-
-    # --------------------------- INFO functions -------------------------------
     def _validate(self):
         validateMsgs = cryosparcValidate()
         if not validateMsgs:
@@ -476,7 +435,7 @@ class ProtCryoSparcRefine3D(ProtRefine3D):
             summary.append("Input Particles: %s" %
                            self.getObjectTag('inputParticles'))
             summary.append("Input Volume: %s" %
-                           self.getObjectTag('referenceVolume'))
+                           self.getObjectTag('refVolume'))
             summary.append("Input Mask: %s" %
                            self.getObjectTag('refMask'))
             summary.append("Symmetry: %s" %
@@ -492,9 +451,6 @@ class ProtCryoSparcRefine3D(ProtRefine3D):
 
     # -------------------------- UTILS functions ------------------------------
 
-    def _getInputParticles(self):
-        return self.inputParticles.get()
-
     def _fillDataFromIter(self, imgSet):
         outImgsFn = self._getFileName('out_particles')
         imgSet.setAlignmentProj()
@@ -507,47 +463,6 @@ class ProtCryoSparcRefine3D(ProtRefine3D):
         createItemMatrix(particle, row, align=ALIGN_PROJ)
         setCryosparcAttributes(particle, row,
                                md.RLN_PARTICLE_RANDOM_SUBSET)
-
-    def _initializeUtilsVariables(self):
-        """
-        Initialize all utils cryoSPARC variables
-        """
-        # Create a cryoSPARC project dir
-        self.projectDirName = getProjectName(self.getProject().getShortName())
-        self.projectPath = pwutils.join(getCryosparcProjectsDir(), self.projectDirName)
-        self.projectDir = createProjectDir(self.projectPath)
-
-    def _initializeCryosparcProject(self):
-        """
-        Initialize the cryoSPARC project and workspace
-        """
-        self._initializeUtilsVariables()
-        # create empty project or load an exists one
-        folderPaths = getProjectPath(self.projectPath)
-        if not folderPaths:
-            self.a = createEmptyProject(self.projectPath, self.projectDirName)
-            self.projectName = self.a[-1].split()[-1]
-        else:
-            self.projectName = str(folderPaths[0])
-
-        self.projectName = String(self.projectName)
-        self._store(self)
-
-        # create empty workspace
-        self.b = createEmptyWorkSpace(self.projectName, self.getRunName(),
-                                      self.getObjComment())
-        self.workSpaceName = String(self.b[-1].split()[-1])
-        self._store(self)
-
-    def _importParticles(self):
-
-        print("Importing Particles")
-
-        # import_particles_star
-        self.importedParticles = doImportParticlesStar(self)
-        self.currenJob = String(self.importedParticles.get())
-        self._store(self)
-        self.par = String(self.importedParticles.get() + '.imported_particles')
 
     def _defineParamsName(self):
         """ Define a list with all protocol parameters names"""
