+++ cryosparc2/protocols/protocol_cryosparc2d.py
@@ -31,12 +31,13 @@ from pwem.protocols import ProtClassify2D
 from pyworkflow.protocol.params import (PointerParam, FloatParam)
 from pyworkflow.utils import replaceExt, createLink
 
+from . import ProtCryosparcBase
 from ..convert import *
 from ..utils import *
 from ..constants import *
 
 
-class ProtCryo2D(ProtClassify2D):
+class ProtCryo2D(ProtCryosparcBase, ProtClassify2D):
     """ Wrapper to CryoSparc 2D clustering program.
         Classify particles into multiple 2D classes to facilitate stack cleaning
         and removal of junk particles. Also useful as a sanity check to
@@ -221,17 +222,6 @@ class ProtCryo2D(ProtClassify2D):
         self._insertFunctionStep('createOutputStep')
 
     # --------------------------- STEPS functions ------------------------------
-    def convertInputStep(self):
-        """ Create the input file in STAR format as expected by Relion.
-        If the input particles comes from Relion, just link the file. 
-        """
-        print(pwutils.greenStr("Importing Particles..."))
-        imgSet = self._getInputParticles()
-        writeSetOfParticles(imgSet, self._getFileName('input_particles'),
-                            self._getTmpPath())
-
-        self._importParticles()
-
     def processStep(self):
         """
         Classify particles into multiples 2D classes
@@ -335,12 +325,6 @@ class ProtCryo2D(ProtClassify2D):
         self._defineOutputs(outputClasses=classes2DSet)
         self._defineSourceRelation(self.inputParticles.get(), classes2DSet)
 
-    def setAborted(self):
-        """ Set the status to aborted and updated the endTime. """
-        ProtClassify2D.setAborted(self)
-        killJob(str(self.projectName.get()), str(self.currenJob.get()))
-        clearJob(str(self.projectName.get()), str(self.currenJob.get()))
-
     # --------------------------- INFO functions -------------------------------
     def _validate(self):
         validateMsgs = cryosparcValidate()
@@ -375,9 +359,6 @@ class ProtCryo2D(ProtClassify2D):
         return [methods]
     
     # --------------------------- UTILS functions ------------------------------
-    def _getInputParticles(self):
-        return self.inputParticles.get()
-
     def _loadClassesInfo(self, filename):
         """ Read some information about the produced 2D classes
         from the metadata file.
@@ -445,48 +426,6 @@ class ProtCryo2D(ProtClassify2D):
             class2Drep = class2D.getRepresentative()
             class2Drep.setLocation(index, fn)
             class2Drep.setSamplingRate(sr)
-
-    def _initializeUtilsVariables(self):
-        """
-        Initialize all utils cryoSPARC variables
-        """
-        # Create a cryoSPARC project dir
-        self.projectDirName = getProjectName(self.getProject().getShortName())
-        self.projectPath = pwutils.join(getCryosparcProjectsDir(),
-                                        self.projectDirName)
-        self.projectDir = createProjectDir(self.projectPath)
-
-    def _initializeCryosparcProject(self):
-        """
-        Initialize the cryoSPARC project and workspace
-        """
-        self._initializeUtilsVariables()
-        # create empty project or load an exists one
-        folderPaths = getProjectPath(self.projectPath)
-        if not folderPaths:
-            self.a = createEmptyProject(self.projectPath, self.projectDirName)
-            self.projectName = self.a[-1].split()[-1]
-        else:
-            self.projectName = str(folderPaths[0])
-
-        self.projectName = String(self.projectName)
-        self._store(self)
-
-        # create empty workspace
-        self.b = createEmptyWorkSpace(self.projectName, self.getRunName(),
-                                      self.getObjComment())
-        self.workSpaceName = String(self.b[-1].split()[-1])
-        self._store(self)
-
-    def _importParticles(self):
-
-        print("Importing Particles...")
-
-        # import_particles_star
-        self.importedParticles = doImportParticlesStar(self)
-        self.currenJob = String(self.importedParticles.get())
-        self._store(self)
-        self.par = String(self.importedParticles.get() + '.imported_particles')
 
     def _defineParamsName(self):
         """ Define a list with all protocol parameters names"""
